Add files via upload
Whoops, I've been a huge fool, Turns out you have to define the "Player" in the trigger as a public gameobject for the thing to work. Anyhow. to make this work, follow the steps to create the desired animation, then apply the respective animation controller on the trigger. Simplicity itself. Anyway, to make up for my lack of activity I will be creating a new component for you guys,